cantina/src/components/userdb-back/requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.83.0
h11==0.13.0
idna==3.3
pydantic==1.10.2
PyYAML==6.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
cryptography
msgpack
